+++ src/test/java/com/playtomic/tests/wallet/repository/WalletRepositoryTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.data.mongo.DataMongoTest;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
@@ -17,7 +18,7 @@ public class WalletRepositoryTest {
 
     @Test
     public void createWalletTest() {
-        Wallet newWallet = new Wallet("John DOe", 0.0);
+        Wallet newWallet = new Wallet("John DOe", BigDecimal.valueOf(0.00));
         walletRepository.save(newWallet);
 
         Optional<Wallet> foundWallet = walletRepository.findById(newWallet.getId());
